Move getClass to separate file
