+++ packages/hierarchy-builder/src/hierarchy-builder/internal/Common.ts
@@ -5,7 +5,7 @@
 
 import naturalCompare from "natural-compare-lite";
 import { assert } from "@itwin/core-bentley";
-import { ECClass, ECSchema, IMetadataProvider } from "../ECMetadata";
+import { ECClass, IMetadataProvider } from "../ECMetadata";
 import {
   HierarchyNodeKey,
   HierarchyNodeLabelGroupingParams,
@@ -15,39 +15,10 @@ import {
   ProcessedCustomHierarchyNode,
   ProcessedInstanceHierarchyNode,
 } from "../HierarchyNode";
-import { getLogger } from "../Logging";
-import { parseFullClassName } from "../Metadata";
+import { getClass } from "./GetClass";
 
 /** @internal */
 export const LOGGING_NAMESPACE = "Presentation.HierarchyBuilder";
-
-/** @internal */
-export async function getClass(metadata: IMetadataProvider, fullClassName: string): Promise<ECClass> {
-  const { schemaName, className } = parseFullClassName(fullClassName);
-  let schema: ECSchema | undefined;
-  try {
-    schema = await metadata.getSchema(schemaName);
-  } catch (e) {
-    assert(e instanceof Error);
-    getLogger().logError(`${LOGGING_NAMESPACE}`, `Failed to get schema "${schemaName} with error ${e.message}."`);
-  }
-  if (!schema) {
-    throw new Error(`Invalid schema "${schemaName}"`);
-  }
-
-  let nodeClass: ECClass | undefined;
-  try {
-    nodeClass = await schema.getClass(className);
-  } catch (e) {
-    assert(e instanceof Error);
-    getLogger().logError(`${LOGGING_NAMESPACE}`, `Failed to get schema "${schemaName} with error ${e.message}."`);
-  }
-  if (!nodeClass) {
-    throw new Error(`Invalid class "${className}" in schema "${schemaName}"`);
-  }
-
-  return nodeClass;
-}
 
 function mergeNodeHandlingParams(
   lhs: InstanceHierarchyNodeProcessingParams | undefined,
@@ -167,4 +138,23 @@ export function mergeSortedArrays<TLhsItem, TRhsItem>(
 /** @internal */
 export function compareNodesByLabel<TLhsNode extends { label: string }, TRhsNode extends { label: string }>(lhs: TLhsNode, rhs: TRhsNode): number {
   return naturalCompare(lhs.label.toLocaleLowerCase(), rhs.label.toLocaleLowerCase());
+}
+
+/** @internal */
+export class BaseClassChecker {
+  private _map = new Map<string, boolean>();
+  private _metadataProvider: IMetadataProvider;
+  public constructor(metadataProvider: IMetadataProvider) {
+    this._metadataProvider = metadataProvider;
+  }
+
+  public async isECClassOfBaseECClass(ecClassNameToCheck: string, baseECClass: ECClass): Promise<boolean> {
+    let isCurrentNodeClassOfBase = this._map.get(`${ecClassNameToCheck}${baseECClass.fullName}`);
+    if (isCurrentNodeClassOfBase === undefined) {
+      const currentNodeECClass = await getClass(this._metadataProvider, ecClassNameToCheck);
+      isCurrentNodeClassOfBase = await currentNodeECClass.is(baseECClass);
+      this._map.set(`${ecClassNameToCheck}${baseECClass.fullName}`, isCurrentNodeClassOfBase);
+    }
+    return isCurrentNodeClassOfBase;
+  }
 }
