+++ packages/hierarchy-builder/src/hierarchy-builder/internal/operators/Grouping.ts
@@ -6,9 +6,8 @@ import { concatMap, from, Observable, of, tap, toArray } from "rxjs";
 import { IMetadataProvider } from "../../ECMetadata";
 import { HierarchyNode, HierarchyNodeKey, ProcessedGroupingHierarchyNode, ProcessedHierarchyNode, ProcessedInstanceHierarchyNode } from "../../HierarchyNode";
 import { getLogger } from "../../Logging";
-import { BaseClassChecker } from "../../Utils";
 import { IPrimitiveValueFormatter } from "../../values/Formatting";
-import { compareNodesByLabel, createOperatorLoggingNamespace, mergeSortedArrays } from "../Common";
+import { BaseClassChecker, compareNodesByLabel, createOperatorLoggingNamespace, mergeSortedArrays } from "../Common";
 import { assignAutoExpand } from "./grouping/AutoExpand";
 import { createBaseClassGroupingHandlers } from "./grouping/BaseClassGrouping";
 import { createClassGroups } from "./grouping/ClassGrouping";
