+++ packages/hierarchy-builder/src/hierarchy-builder/queries/InstanceLabelSelectClauseFactory.ts
@@ -4,7 +4,7 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { IMetadataProvider } from "../ECMetadata";
-import { getClass } from "../internal/Common";
+import { getClass } from "../internal/GetClass";
 import { createConcatenatedValueJsonSelector, createRawPropertyValueSelector, TypedValueSelectClauseProps } from "./ecsql-snippets/ECSqlValueSelectorSnippets";
 
 /**
